Update CacheValue hashcode and equality check to rely only on key
While technically imprecise, this is a developer class that should be used with care. In general, only comparisons to other instances of the same type and from the same RDD should be made, in which case it's generally safe to assume there's a unique output key.

Initial motivation for making this change is because at least one CacheValue consists of a data type that is not easily supported by the default equals method - Array (IIRC - PreShuffle). While we could implement an explicit equals method for the affected classes, it's technically faster to compare only by keys anyways, so this is also a performance improvement.