Add latency statistics computation for post-shuffle operations. Update pre-shuffle operation to use the same LineageManager.finalize approach.
In LineageManager.finalizeTaskCache, agg stats will be uploaded if the rdd is an instance of the recently defined LatencyStatsTap trait.

ShuffleMapTask (preshuffle) agg stats collection remains the same, with the exception that the resulting stats are attached to the rdd before finalizing it via LineageManager. Previously, the task itself did the uploading (but this resulted in a non-centralized upload locations).

ShuffledLRDD and CoGroupedLRDD (postshuffle operations) both override the TapRDD.compute function. They measure the time it takes to normally compute the rdd's iterator as well as the time it takes to iterate through the entire iterator. It leverages a CompletionIterator (basically an iterator with a final callback) to compute aggregate latency statistics and assign them to the RDD once exhausted. (Dev note: both compute methods are near-identical and could probably be refactored accordingly).