Update LineageWrapper to store precomputed number of partitions for the tapRDD. Update PerfTraceCalculator to leverage this field where necessary.
Previously had issues in external session where some RDD.getPartitions implementations relied on information no longer present (due to serialization or session close).

One example of this is TapRDD - it relies on firstParent.getPartitions. If the parent is a HadoopRDD, the hadoop RDD will attempt to load its job conf from a broadcast variable through a spark context that is no longer active. Thus, we compute the number of partitions beforehand and save them in LineageCacheDependencies.