Update SlowestInputsCalculator to provide option to skip backwards trace
Initial lineage trace (from outputs to relevant inputs) can be expensive and suboptimal. In some cases it's faster to simply use every single input (a guaranteed superset of inputs that lead to the final output) and skip the backwards/lineage trace entirely. Doing so means we would start right at the forward trace.